requirements.txt
transforms3d
smilelogging==0.2.10
imageio-ffmpeg
lmdb==1.2.1
tqdm==4.61.2
ConfigArgParse==1.5.1
dill==0.3.4
ipdb==0.13.9
numpy==1.21
ptflops==0.6.6
imageio==2.9.0
opencv_python==4.5.3.56
pandas==1.3.0
torchvision==0.10.0
matplotlib==3.4.2
onnxruntime==1.8.1
onnx==1.10.1
pytest==6.2.4
Pillow==9.0.1
PyYAML==5.4.1
scikit_learn==0.24.2
uutils==0.0.3
lpips
